requirements.txt
# Function dependencies, for example:
# package>=version
transformers[torch]==4.5.1
sentencepiece==0.1.95
google-cloud-firestore